+++ client/src/core/todo/todo.util.ts
@@ -1,0 +1,2 @@
+export const onlyDate = (date: Date): Date => new Date(date.getFullYear(), date.getMonth(), date.getDate());
+export const isEqualDate = (d1: Date, d2: Date): boolean => onlyDate(d1).getTime() === onlyDate(d2).getTime();
