+++ client/vite.config.ts
@@ -10,6 +10,7 @@ export default defineConfig({
       '@components': resolve(__dirname, './src/components'),
       '@container': resolve(__dirname, './src/container'),
       '@core': resolve(__dirname, './src/core'),
+      '@todo': resolve(__dirname, './src/core/todo'),
       '@images': resolve(__dirname, './src/images'),
       '@page': resolve(__dirname, './src/page'),
       '@util': resolve(__dirname, './src/util'),
